fixed hardcoded stripe to real data
